Complete API renaming into DPS. Add sample local_settings
