+++ bite/dps/apps.py
@@ -20,5 +20,5 @@
 from django.apps import AppConfig
 
 
-class ApiConfig(AppConfig):
-    name = 'api'
+class DPSConfig(AppConfig):
+    name = 'dps'
