+++ bite/telemetry/management/commands/dispatcher.py
@@ -31,7 +31,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 from django.core.exceptions import ObjectDoesNotExist
 
-from api.models import Device
+from dps.models import Device
 
 
 class Command(BaseCommand):
